requirements.txt
numpy
matplotlib
streamlit
plotly
google-generativeai
